Added step file for Cypher compatibility suite. Also refactured test code
