--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "neokit"]
-	path = neokit
-	url = https://github.com/nigelsmall/neokit.git
