+++ runtests.sh
@@ -62,8 +62,7 @@ echo "Running tests with $(python --version)"
 pip install --upgrade -r ${DRIVER_HOME}/test_requirements.txt
 echo ""
 TEST_RUNNER="coverage run -m ${UNITTEST} discover -vfs ${TEST}"
-EXAMPLES_RUNNER="coverage run -m ${UNITTEST} discover -vfs examples"
-BEHAVE_RUNNER="behave test/tck"
+BEHAVE_RUNNER="behave --tags=-db,-in_dev test/tck"
 if [ ${RUNNING} -eq 1 ]
 then
     ${TEST_RUNNER}
